requirements.txt
-e git+git@github.com:assimilate-dev/appyrition.git@57f047ce9103e9aa58608bcf76e6e39e4ec88e5d#egg=appyrition
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.2.1
idna==2.10
Markdown==3.3.3
pycparser==2.20
PyJWT==2.0.0
requests==2.25.1
six==1.15.0
urllib3==1.26.2
